Define and document common compiler extensions
We have defined the commonly used compiler extensions as a set of
macros. These macros are available when GCC and Clang are used; on other
compilers these macros degrade to no-ops with suitable compiler
warnings.

We have also prepared comprehensive documentation for these macros as
`m4` templates. These templates are used to conditionally generate the
`pandoc` flavoured Markdown files that are subsequently used to generate
the man pages.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>