Rename to `libchrysalid`
The name `libchrysalis` already exists as a repository in GitHub; in
order to avoid the name conflict we have renamed to `libchrysalid`.
Since the word `chrysalid` is related to `chrysalis` we do not lose out
on the original meaning and rationale behind the choice of name.

Accordingly, all code assests which referred to `libchrysalis` now refer
to `libchrysalid` instead. The main changes have been in the
documentation; for the code, there has been no change in the existing
`c_` namespace prefix.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>