Generate man page documentation for hptr interface
The man page documentation for the heap pointer interface is now being
generated from the source code comments marked with special tags. A
combination of m4 macros and shell scripts is used to prepare the man
page along with the pandoc tool.

Currently, we are documenting the following aspects of the heap pointer
interface:
  - Types
  - Constants
  - Macros
  - Functions
  - Test cases

The shell scripts responsible for generating the man page are still
rough and need to be improved; this will be done incrementally as the
other components of libchrysalid are developed.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>