Add basic UTF-8 string functionality
Basic support for UTF-8 strings has been added through the `cy_utf8_t`
type and its supporting macros and functions. The UTF-8 strings are
reference counted, and we are able to perform both shallow and deep
copies on them. We also have support for regular expression pattern
matching.

The string interface relies on the external PCRE2 and utf8.h libraries.
The utf8.h header library is available as a Git submodule; PCRE2 is
required to be installed in the host system.

The string interface is reasonably well-tested and documented; however,
the man-page documentation still needs to be generated, as does
commenting the tests. These have been left for a future commit in favour
of continued development.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>