Implement JSON interface
The JSON interface has been implemented as a wrapper around the
cJSON library. Currently, the interface supports parsing a string
to a JSON data structure along with basic accessors. Mutators have
not been implemented, and may be done in future if the need arises.

The interface has been released to the `main` branch early in order
to allow the continued development of client libraries. Although a
sanity check in the form of an example application has been done,
formal testing is still pending, as is complete documentation. These
are expected to be completed in the subsequent commits.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>