Add log manager
Support for logging has been introduced. We are now able to log messages
using the syslog service. The log interface primarily consists of macros
that write formatted log messages with a given priority level.

Testing and documentation are still pending; however, we have released
the code early to the `main` branch in order to allow continued
development of client code.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>