Introduce config interface and other improvements
The config interface has been developed based on the external
`iniparser` submodule. As of now, we have basic support for parsing ini
files, but will include support in future for TOML and Lua configuration
files (if required).

At the same time, several small improvements have been introduced based
on the experience of using this library with another library
(`libelmy`). The improvements are as follows:

  - Addition of `CY_SAFE` and `CY_WEAK` macros
  - Fixing of memory leaks in the `cy_json_t` interface
  - Addition of log facility and severity enumerations
  - Addition of `cy_utf8_new_fmt()`
  - Addition of `cy_utf8_replace()`
  - Addition of `cy_utf8_escape_json()`
  - Better build targets

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>